+++ samples/porting/fal_flash_sfud_port.c
@@ -13,14 +13,14 @@
 
 extern sfud_flash sfud_norflash0;
 
-static int read(uint32_t offset, uint8_t *buf, size_t size)
+static int read(long offset, uint8_t *buf, size_t size)
 {
     sfud_read(&sfud_norflash0, nor_flash0.addr + offset, size, buf);
 
     return size;
 }
 
-static int write(uint32_t offset, const uint8_t *buf, size_t size)
+static int write(long offset, const uint8_t *buf, size_t size)
 {
     if (sfud_write(&sfud_norflash0, nor_flash0.addr + offset, size, buf) != SFUD_SUCCESS)
     {
@@ -30,7 +30,7 @@ static int write(uint32_t offset, const uint8_t *buf, size_t size)
     return size;
 }
 
-static int erase(uint32_t offset, size_t size)
+static int erase(long offset, size_t size)
 {
     if (sfud_erase(&sfud_norflash0, nor_flash0.addr + offset, size) != SFUD_SUCCESS)
     {
